+++ src/graphql/network.rs
@@ -266,8 +266,9 @@ impl Network {
         let codec = bincode::DefaultOptions::new();
         for &id in &self.inner.customer_ids {
             #[allow(clippy::cast_sign_loss)] // u32 stored as i32 in database
-            let Some(value) = map.get_by_id(id)? else {
-                continue
+            let Some(value) = map.get_by_id(id)?
+            else {
+                continue;
             };
             let customer = codec
                 .deserialize::<database::Customer>(value.as_ref())