+++ src/graphql/node/crud.rs
@@ -50,7 +50,7 @@ impl NodeQuery {
         let store = crate::graphql::get_store(ctx).await?;
         let map = store.node_map();
         let Some(value) = map.get_by_id(i)? else {
-           return Err("no such node".into())
+            return Err("no such node".into());
         };
         Ok(bincode::DefaultOptions::new()
             .deserialize(value.as_ref())